move bls functions to blsutils.go
