+++ toolkit/tools/pkg/imagecustomizerlib/customizeuki.go
@@ -773,148 +773,6 @@ func readKernelCmdlinesFromGrubCfg(bootDir string, grubCfgRelPath string) (map[s
 	return formatKernelCmdlineFromLinuxLines(linuxLines)
 }
 
-// parseBLSLinuxValue parses the value following a "linux" key in a BLS entry.
-func parseBLSLinuxValue(lineTokens []grub.Token) (string, error) {
-	if len(lineTokens) != 2 {
-		return "", fmt.Errorf("expected 1 value token, found %d", len(lineTokens)-1)
-	}
-	valueToken := lineTokens[1]
-	if valueToken.Type != grub.WORD {
-		return "", fmt.Errorf("unexpected non-word value token: %v", valueToken.RawContent)
-	}
-	if len(valueToken.SubWords) != 1 {
-		return "", fmt.Errorf("expected 1 subword, found %d", len(valueToken.SubWords))
-	}
-	sw := valueToken.SubWords[0]
-	if sw.Type != grub.KEYWORD_STRING && sw.Type != grub.STRING {
-		return "", fmt.Errorf("unexpected non-string subword: %v", sw.RawContent)
-	}
-	return sw.Value, nil
-}
-
-// parseBLSTitleValue parses the value following a "title" key in a BLS entry.
-func parseBLSTitleValue(lineTokens []grub.Token) (string, error) {
-	if len(lineTokens) < 2 {
-		return "", fmt.Errorf("expected at least 1 value token, found 0")
-	}
-	valueParts := make([]string, 0, len(lineTokens)-1)
-	for _, valueToken := range lineTokens[1:] {
-		if valueToken.Type != grub.WORD {
-			return "", fmt.Errorf("unexpected non-word value token: %v", valueToken.RawContent)
-		}
-		if len(valueToken.SubWords) < 1 {
-			return "", fmt.Errorf("expected at least 1 subword, found 0")
-		}
-		var sb strings.Builder
-		for _, sw := range valueToken.SubWords {
-			if sw.Type != grub.KEYWORD_STRING && sw.Type != grub.STRING {
-				return "", fmt.Errorf("unexpected non-string subword: %v", sw.RawContent)
-			}
-			sb.WriteString(sw.Value)
-		}
-		valueParts = append(valueParts, sb.String())
-	}
-	return strings.Join(valueParts, " "), nil
-}
-
-// readKernelCmdlinesFromBLSEntries reads Boot Loader Specification (BLS) entries in {bootDir}/loader/entries/*.conf,
-// extracting a kernel-to-cmdline mapping for non-recovery entries.
-func readKernelCmdlinesFromBLSEntries(bootDir string) (map[string][]grubConfigLinuxArg, error) {
-	entriesDir := filepath.Join(bootDir, "loader", "entries")
-	entries, err := os.ReadDir(entriesDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read BLS entries directory (%s):\n%w", entriesDir, err)
-	}
-
-	kernelToArgs := make(map[string][]grubConfigLinuxArg)
-	for _, entry := range entries {
-		if entry.IsDir() || !strings.HasSuffix(entry.Name(), ".conf") {
-			logger.Log.Debugf("Skipping non-.conf BLS entry file (%s) in directory (%s)", entry.Name(), entriesDir)
-			continue
-		}
-
-		absPath := filepath.Join(entriesDir, entry.Name())
-		content, err := os.ReadFile(absPath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to read BLS entry file (%s):\n%w", absPath, err)
-		}
-
-		// BLS entries are grub-syntax key/value pairs, so we can leverage the grub parsing code here.
-		tokens, err := grub.TokenizeConfig(string(content))
-		if err != nil {
-			return nil, fmt.Errorf("failed to tokenize BLS entry (%s):\n%w", absPath, err)
-		}
-		lines := grub.SplitTokensIntoLines(tokens)
-
-		var linux string
-		var title string
-		var args []grubConfigLinuxArg
-
-		for _, line := range lines {
-			if len(line.Tokens) == 0 {
-				continue
-			}
-
-			keyToken := line.Tokens[0]
-			if keyToken.Type != grub.WORD {
-				return nil, fmt.Errorf("unexpected non-word token in BLS entry (%s): %v", absPath, keyToken.RawContent)
-			}
-			if len(keyToken.SubWords) != 1 {
-				return nil, fmt.Errorf("unexpected token with multiple subwords in BLS entry (%s): %v", absPath, keyToken.RawContent)
-			}
-			if keyToken.SubWords[0].Type != grub.KEYWORD_STRING {
-				return nil, fmt.Errorf("unexpected non-string token in BLS entry (%s): %v", absPath, keyToken.SubWords[0].RawContent)
-			}
-			key := keyToken.SubWords[0].Value
-
-			switch key {
-			case "linux":
-				if linux != "" {
-					return nil, fmt.Errorf("duplicate key (%s) in BLS entry (%s)", key, absPath)
-				}
-				linuxValue, err := parseBLSLinuxValue(line.Tokens)
-				if err != nil {
-					return nil, fmt.Errorf("failed to parse BLS key (%s) for entry (%s):\n%w", key, absPath, err)
-				}
-				linux = filepath.Base(linuxValue)
-			case "title":
-				titleValue, err := parseBLSTitleValue(line.Tokens)
-				if err != nil {
-					return nil, fmt.Errorf("failed to parse BLS key (%s) for entry (%s):\n%w", key, absPath, err)
-				}
-				title = titleValue
-			case "efi", "uki", "uki-url":
-				return nil, fmt.Errorf("BLS entry (%s) uses '%s' key, which is not supported", absPath, key)
-			case "options":
-				// "options" may appear multiple times per BLS spec.
-				lineArgs, err := ParseCommandLineArgs(line.Tokens[1:])
-				if err != nil {
-					return nil, fmt.Errorf("failed to parse BLS key (%s) for entry (%s):\n%w", key, absPath, err)
-				}
-				args = append(args, lineArgs...)
-			}
-		}
-
-		if linux == "" {
-			return nil, fmt.Errorf("BLS entry (%s) is missing 'linux' key", absPath)
-		}
-
-		// Entries without titles are treated as normal entries.
-		if isRecoveryOrRescueTitle(title) {
-			logger.Log.Debugf("Skipping recovery/rescue BLS entry with title (%s) in file (%s)", title, absPath)
-			continue
-		}
-
-		if _, exists := kernelToArgs[linux]; exists {
-			return nil, fmt.Errorf("duplicate BLS entries for kernel (%s) in (%s)", linux, entriesDir)
-		}
-
-		kernelToArgs[linux] = args
-	}
-
-	return kernelToArgs, nil
-}
-
 // readNonRecoveryKernelCmdlinesFromGrubCfg reads the first non-recovery kernel's command-line
 // arguments from grub.cfg. This reproduces the original osmodifier behavior for non-BLS distros.
 func readNonRecoveryKernelCmdlinesFromGrubCfg(grubCfgPath string, argNames []string) (map[string]string, error) {
@@ -944,25 +802,6 @@ func readNonRecoveryKernelCmdlinesFromGrubCfg(grubCfgPath string, argNames []str
 	}
 
 	return filterKernelArgsByName(args, argNames), nil
-}
-
-// readNonRecoveryKernelCmdlinesFromBLS reads the first non-recovery kernel's command-line
-// arguments from BLS entry files.
-func readNonRecoveryKernelCmdlinesFromBLS(bootDir string, argNames []string) (map[string]string, error) {
-	kernelToArgs, err := readKernelCmdlinesFromBLSEntries(bootDir)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(kernelToArgs) > 1 {
-		return nil, fmt.Errorf("expected 1 non-recovery BLS entry, found %d", len(kernelToArgs))
-	}
-
-	for _, args := range kernelToArgs {
-		return filterKernelArgsByName(args, argNames), nil
-	}
-
-	return nil, fmt.Errorf("no non-recovery BLS entries found")
 }
 
 // grubKernelArgsToStringMap converts a kernel-to-args mapping from the parsed grubConfigLinuxArg format to a simple
